AdaBoost: Move fitter to modelling
