+++ Orange/modelling/__init__.py
@@ -2,3 +2,4 @@ from .base import *
 
 from .knn import *
 from .tree import *
+from .ada_boost import *
